Migrate formatDateForDateRage and displayDateRange over to utils/formatters.js
