+++ src/utils/formatters.js
@@ -39,6 +39,37 @@ export function parseSwaggerDate(dateString) {
   return undefined;
 }
 
+// Format date for display of dates summaries
+const formatDateForDateRange = (date, formatType) => {
+  let format = '';
+  switch (formatType) {
+    case 'long':
+      format = 'ddd, MMM DD';
+      break;
+    case 'condensed':
+      format = 'MMM D';
+      break;
+    default:
+      format = 'ddd, MMM DD';
+  }
+  if (date) {
+    return moment(date).format(format);
+  }
+  return undefined;
+};
+
+export const displayDateRange = (dates, formatType = 'long') => {
+  let span = '';
+  let firstDate = '';
+  if (dates.length > 1) {
+    span = ` - ${formatDateForDateRange(dates[dates.length - 1], formatType)}`;
+  }
+  if (dates.length >= 1) {
+    firstDate = formatDateForDateRange(dates[0], formatType);
+  }
+  return firstDate + span;
+};
+
 // Format orders type (ex: PERMANENT_CHANGE_OF_STATION => Permanent change of station)
 export function formatOrderType(orderType) {
   return orderType
